NN-5925 refactoring, ensure punishment removal removes activated links
