+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/controllers/reported/PunishmentsController.kt
@@ -5,7 +5,6 @@ import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.http.HttpStatus
 import org.springframework.security.access.prepost.PreAuthorize
-import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
@@ -14,12 +13,10 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.config.ErrorResponse
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.AdditionalDaysDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.SuspendedPunishmentDto
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.PrivilegeType
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.PunishmentType
-import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReasonForChange
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.AdjudicationDomainEventType
 import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.reported.PunishmentsService
 import java.time.LocalDate
@@ -56,16 +53,6 @@ data class PunishmentRequest(
   val consecutiveChargeNumber: String? = null,
   @Schema(description = "optional amount - money being recovered for damages - only use if type is DAMAGED_OWED")
   val damagesOwedAmount: Double? = null,
-)
-
-@Schema(description = "punishment comment request")
-data class PunishmentCommentRequest(
-  @Schema(description = "id of punishment comment")
-  val id: Long? = null,
-  @Schema(description = "punishment comment")
-  val comment: String,
-  @Schema(description = "punishment reason for change")
-  val reasonForChange: ReasonForChange? = null,
 )
 
 @Schema(description = "active punishment dto")
@@ -174,96 +161,4 @@ class PunishmentsController(
     @PathVariable(name = "offenderBookingId") offenderBookingId: Long,
   ): List<ActivePunishmentDto> =
     punishmentsService.getActivePunishments(offenderBookingId = offenderBookingId)
-
-  @PreAuthorize("hasRole('ADJUDICATIONS_REVIEWER') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "create punishment comment",
-    responses = [
-      io.swagger.v3.oas.annotations.responses.ApiResponse(
-        responseCode = "201",
-        description = "Punishment comment created",
-      ),
-      io.swagger.v3.oas.annotations.responses.ApiResponse(
-        responseCode = "415",
-        description = "Not able to process the request because the payload is in a format not supported by this endpoint.",
-        content = [
-          io.swagger.v3.oas.annotations.media.Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = ErrorResponse::class),
-          ),
-        ],
-      ),
-    ],
-  )
-  @PostMapping(value = ["/{chargeNumber}/punishments/comment"])
-  @ResponseStatus(HttpStatus.CREATED)
-  fun createPunishmentComment(
-    @PathVariable(name = "chargeNumber") chargeNumber: String,
-    @RequestBody punishmentCommentRequest: PunishmentCommentRequest,
-  ): ReportedAdjudicationResponse {
-    val reportedAdjudication = punishmentsService.createPunishmentComment(
-      chargeNumber = chargeNumber,
-      punishmentComment = punishmentCommentRequest,
-    )
-
-    return ReportedAdjudicationResponse(reportedAdjudication)
-  }
-
-  @PreAuthorize("hasRole('ADJUDICATIONS_REVIEWER') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "update punishment comment",
-    responses = [
-      io.swagger.v3.oas.annotations.responses.ApiResponse(
-        responseCode = "200",
-        description = "Punishment comment updated",
-      ),
-      io.swagger.v3.oas.annotations.responses.ApiResponse(
-        responseCode = "415",
-        description = "Not able to process the request because the payload is in a format not supported by this endpoint.",
-        content = [
-          io.swagger.v3.oas.annotations.media.Content(
-            mediaType = "application/json",
-            schema = Schema(implementation = ErrorResponse::class),
-          ),
-        ],
-      ),
-    ],
-  )
-  @PutMapping(value = ["/{chargeNumber}/punishments/comment"])
-  @ResponseStatus(HttpStatus.OK)
-  fun updatePunishmentComment(
-    @PathVariable(name = "chargeNumber") chargeNumber: String,
-    @RequestBody punishmentCommentRequest: PunishmentCommentRequest,
-  ): ReportedAdjudicationResponse {
-    val reportedAdjudication = punishmentsService.updatePunishmentComment(
-      chargeNumber = chargeNumber,
-      punishmentComment = punishmentCommentRequest,
-    )
-
-    return ReportedAdjudicationResponse(reportedAdjudication)
-  }
-
-  @PreAuthorize("hasRole('ADJUDICATIONS_REVIEWER') and hasAuthority('SCOPE_write')")
-  @Operation(
-    summary = "delete punishment comment",
-    responses = [
-      io.swagger.v3.oas.annotations.responses.ApiResponse(
-        responseCode = "200",
-        description = "Punishment comment deleted",
-      ),
-    ],
-  )
-  @DeleteMapping(value = ["/{chargeNumber}/punishments/comment/{punishmentCommentId}"])
-  @ResponseStatus(HttpStatus.OK)
-  fun deletePunishmentComment(
-    @PathVariable(name = "chargeNumber") chargeNumber: String,
-    @PathVariable(name = "punishmentCommentId") punishmentCommentId: Long,
-  ): ReportedAdjudicationResponse {
-    val reportedAdjudication = punishmentsService.deletePunishmentComment(
-      chargeNumber = chargeNumber,
-      punishmentCommentId = punishmentCommentId,
-    )
-
-    return ReportedAdjudicationResponse(reportedAdjudication)
-  }
 }
