+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/entities/ReportedAdjudication.kt
@@ -145,6 +145,10 @@ data class ReportedAdjudication(
 
   fun getPunishments() = this.punishments.filter { it.deleted != true }
 
+  fun removeActivatedByLink(activatedFrom: String) {
+    this.getPunishments().filter { p -> p.activatedByChargeNumber == activatedFrom }.forEach { p -> p.activatedByChargeNumber = null }
+  }
+
   fun addPunishment(punishment: Punishment) = this.punishments.add(punishment)
 
   fun clearPunishments() = this.punishments.clear()
